+++ libraries/manager_client.rb
@@ -10,10 +10,9 @@ class Chef
     class ManagerClient
       # Communicates with the running Tomcat container at a given port and returns
       # an array of application names currently deployed on the server. If a
-      # communication error occurs, an exception is thrown
+      # communication error occurs, an exception is thrown.
       def self.get_deployed_applications(port, tomcat_script_pass)
         deployed_apps = []
-        success = false
         begin
           open(
             "http://127.0.0.1:#{port}/manager/text/list",
@@ -23,12 +22,13 @@ class Chef
             # OK - Listed applications for virtual host localhost
             # /probe:running:1:probe
             # /manager:running:0:manager
-            response_arr = f.read.split(/\n/)
+            response = f.read
+            response_arr = response.split(/\n/)
             response_arr.each_with_index do |r, i|
               if i.zero?
-                success = r[0, 2] == 'OK'
+                raise response unless r[0, 2] == 'OK'
               elsif r[0, 1] == '/'
-                deployed_apps.push(r.split(':')[-1, 1])
+                deployed_apps.push(r.split(':')[-1, 1][0])
               end
             end
           end
@@ -37,6 +37,27 @@ class Chef
           Chef::Log.error "Error occurred when communicating with Tomcat server: #{e}"
           raise
         end
+      end
+
+      # Communicates with the running Tomcat container at a given port and attempts
+      # to undeploy an application. If a communication error occurs, or the application
+      # cannot be undeployed for some reason, an exception is thrown. Otherwise,
+      # the server response is returned
+      def self.undeploy_application(port, tomcat_script_pass, application_name)
+        puts "HERE!!!!"
+        open(
+          "http://127.0.0.1:#{port}/manager/text/undeploy?path=/#{application_name}",
+          http_basic_authentication: ['tomcat-script', tomcat_script_pass.to_s]
+        ) do |f|
+          # First result will be the response for this command
+          response = f.read
+          status = response.split(' ')[0]
+          raise response unless status == 'OK'
+          return response
+        end
+      rescue => e
+        Chef::Log.error "Error occurred when communicating with Tomcat server: #{e}"
+        raise
       end
     end
   end